\name{pmcgd-package}
\alias{pmcgd-package}
\alias{pmcgd}
\docType{package}
\title{pmcgd - Parsimonious Mixtures of Contaminated Gaussian Distributions}
\description{
This package allows for fitting, according to the expectation-conditional maximization algorithm, of the family of 14 parsimonious mixtures of contaminated Gaussian distributions discussed in Punzo & McNicholas (2013).
Some likelihood-based model selection criteria can be adopted to select the best model in the family and the best number of mixture components.
}
\details{
\tabular{ll}{
Package: \tab pmcgd\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-10\cr
License: \tab GNU-2\cr
}
}
\references{
%% Ingrassia, S., Minotti, S. C., and Vittadini, G. (2012). Local statistical modeling via the cluster-weighted approach with elliptical distributions. \emph{Journal of Classification}, \bold{29}(3), 363-401.

Punzo, A., and McNicholas, P. D. (2013). Outlier Detection via Parsimonious Mixtures of Contaminated Gaussian Distributions. \emph{arXiv.org} e-print \bold{1305.4669}, available at: \url{http://arxiv.org/abs/1305.4669}.

}
\author{
Punzo A., McNicholas, P. D.

Maintainer: Punzo Antonio <antonio.punzo@unict.it>
}
%%\note{
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{MS}},\code{\link{class}}
}